AVDSBack/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-multipart==0.0.9
djangorestframework-camel-case==1.4.2
mssql-django==1.3
requests
Pillow
